% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMapping.R
\name{acession.convert}
\alias{acession.convert}
\title{Map uniprot ID to HGNC symbol}
\usage{
acession.convert(mat, verbose = T)
}
\arguments{
\item{mat}{an acession.matrix}
}
\description{
Input an acession matrix to subset it accordingly. It will
map uniprot IDs to their corresponding HGNC symbols.
}
\seealso{
Other id: 
\code{\link{acession.matrix}()}
}
\author{
flassen
}
\concept{id}
