% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMapping.R
\name{acession.matrix}
\alias{acession.matrix}
\title{Split acession ID into matrix}
\usage{
acession.matrix(vec)
}
\description{
Generase an acession ID matrix that
can be further subsetted according to isoforms, gene symbol
and species.
}
\seealso{
Other id: 
\code{\link{acession.convert}()}
}
\author{
flassen
}
\concept{id}
