% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe data}
\usage{
describe(data, checkfor = NULL, verbose = F)
}
\arguments{
\item{data}{a data.frame with proteomics data}
}
\description{
A function that will take a data.frame
and describe the columns, type of data which will be returned in
a list
}
\author{
flassen
}
