% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designate.R
\name{designate}
\alias{designate}
\title{Designate significant observations}
\usage{
designate(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{any arguments that can be inputted into dplyr::filter.}
}
\description{
Designate observations as significant based on filtering them using dplyr::filter.
}
\examples{
\dontrun{
data = df \%>\% genoppi.ttest() ∂\%>\% designate(FDR < 0.1, pvalue < 0.05)}
}
\seealso{
Other genoppi: 
\code{\link{enrichment}()},
\code{\link{genoppi.database}()},
\code{\link{genoppi}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotScatter}()},
\code{\link{plotVolcano}()}
}
\author{
frhl
}
\concept{genoppi}
