% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Detect protein and gene IDs}
\usage{
detect(data, entry, details = F)
}
\arguments{
\item{data}{a data.frame}

\item{entry}{a string or regular expression.}

\item{details}{will return the entire data.frame which contains the entry}
}
\value{
row-wise boolean indiciating whether the entry is present in the data.
}
\description{
Look up HGNC/UNIPROT aliases to see whether an entry can be found.
}
\note{

}
\seealso{
Other id: 
\code{\link{acession.convert}()},
\code{\link{acession.matrix}()}
}
\author{
flassen
}
\concept{id}
