% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Calculate enrichment using Fisher exact test}
\usage{
enrichment(data, bait, reference)
}
\arguments{
\item{data}{data.frame containing columns gene and significance}

\item{bait}{designated bait in the data.}

\item{reference}{data base reference}
}
\description{
Calculate enrichment of a data and reference data base for
a specific bait and its interactors. Will do p-test using a fisher exact test
based on the overlap between the main data and the reference data which both
interacts with bait.
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{genoppi.database}()},
\code{\link{genoppi}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotScatter}()},
\code{\link{plotVolcano}()}
}
\concept{genoppi}
