% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandProxies.R
\name{expandProxies}
\alias{expandProxies}
\title{Expand proxy interactions}
\usage{
expandProxies(data, cutoff.fdr = 0.1, top = 10, verbose = T)
}
\value{
a list of proteins that has inweb interactors in the dataset.

a table that can be inputted to genoppi
}
\description{
Somtimes in Proteomics, the antibody is specific to another protein other
than the bait. This script will look throught the top enriched proteins (i.e. logFC >> 0),
and see whether they have interactors found in inweb and the current dataset.
}
