% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesData.R
\docType{data}
\name{genes_snps}
\alias{genes_snps}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{hash} of length 18116.}
\usage{
data(snps_genes)
}
\description{
data from genoppi used to map snps to gene. 
Todo: ask april about reference, think its hg19
}
\examples{
data(genes_snps)

}
\keyword{datasets}
