% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTable.R
\name{geneTable}
\alias{geneTable}
\title{table genes and snps}
\usage{
geneTable(snpgene, outfile = NULL, delim = ";")
}
\arguments{
\item{snpgene}{a list named by genes containing snps, i.e. the result generated by snpToGene}
}
\description{
creates a 2xn table with genes in the first
column and the associated snps in the second column.
}
\examples{
\dontrun{geneTable(snpToGene('rs142920272'))}
}
\author{
flassen
}
