% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppiMeta.R
\name{genoppi.database}
\alias{genoppi.database}
\title{Genoppi databases}
\usage{
genoppi.database(bait, reference = "inweb")
}
\arguments{
\item{bait}{The bait that is to quries within the reference}

\item{reference}{the reference database. Currently only supports InWeb.}
}
\description{
Genoppi databases
}
