% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppiMeta.R
\name{genoppi.enrichment}
\alias{genoppi.enrichment}
\title{Calculate enrichment using Fisher exact test}
\usage{
genoppi.enrichment(data, bait, reference)
}
\arguments{
\item{data}{data.frame containing columns gene and significance}

\item{bait}{designated bait in the data.}

\item{reference}{data base reference}
}
\description{
Calculate enrichment using Fisher exact test
}
