% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppi.R
\name{genoppi}
\alias{genoppi}
\title{Genoppi analysis pipeline}
\usage{
genoppi(
  args,
  input_file,
  fc_cutoff,
  fdr_cutoff,
  p_cutoff,
  bait_name,
  imp_list_file,
  gene_lists_file,
  output_stats_file,
  output_plots_file
)
}
\arguments{
\item{input_file}{Genoppi input table}

\item{fc_cutoff}{character "positive", "negative", or "both"}

\item{fdr_cutoff}{numeric value or "NA"}

\item{p_cutoff}{numeric value or "NA"}

\item{bait_name}{bait name}

\item{gene_lists_file}{file containing gene list file names, or "NA"}

\item{output_stats_file}{output file containing moderated t-test stats}

\item{output_plots_file}{output file containing plots}

\item{description}{the genoppi combined analysis script. This will run the pipeline
generating correlation plots,}
}
\description{
Genoppi analysis pipeline
}
\author{
aprilkim/Yuhanshu/flassen
}
