% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppiMeta.R
\name{genoppi.significant}
\alias{genoppi.significant}
\title{Genoppi significanct}
\usage{
genoppi.significant(data, ...)
}
\description{
Based on the dplyr::filter function, used
to designate what proteins are significant or not.
}
\examples{
\dontrun{
data = df \%>\% genoppi.ttest()
data \%>\% genoppi.significant(FDR < 0.1, pvalue < 0.05)}
}
\author{
frhl
}
