% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderatedTTest.R
\name{genoppi.ttest}
\alias{genoppi.ttest}
\title{Calculate a moderated t-test}
\usage{
genoppi.ttest(df, keep = c("imputed"))
}
\arguments{
\item{df}{a data.frame with 'rep' in column names.}

\item{keep}{a vector of characters containing colnames that should be kept in the data.frame}
}
\description{
alculate a moderated t-test
}
\author{
April Kim Frederik Heymann
}
