% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactors.R
\name{interactors}
\alias{interactors}
\title{Genoppi databases of interactors}
\usage{
interactors(bait, verbose = F, reference = "inweb")
}
\arguments{
\item{bait}{The bait that is to quries within the reference}

\item{verbose}{how many interactors were found in the database?}

\item{reference}{the reference database. Currently only supports InWeb.}
}
\value{
a data.frame with genename and significant, that indicates whether
the bait-database interactor exists.
}
\description{
Input a bait and the current interactors found in the selected
database will return a list
}
\examples{
\dontrun{d1 <- interactors('CACNA1C')}
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{enrichment}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotScatter}()},
\code{\link{plotVolcano}()}
}
\concept{genoppi}
