% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inwebData.R
\docType{data}
\name{inweb_hash}
\alias{inweb_hash}
\title{InWeb Hash Table}
\format{An object of class \code{hash} of length 18478.}
\usage{
inweb_hash
}
\description{
InWeb Hash table downloaded on october 2018.
}
\keyword{datasets}
