% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFC.R
\name{logFC}
\alias{logFC}
\title{Transform data}
\usage{
logFC(df)
}
\arguments{
\item{df}{a data.frame with bait and controls}
}
\description{
Calculates log fold change by substracting
a case column with a bait column
}
\author{
flassen
}
