% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapp.R
\name{mapp}
\alias{mapp}
\title{map different gene names to one single gene name}
\usage{
mapp(vec, type = "gene", verbose = T)
}
\arguments{
\item{type}{protein or gene?}

\item{verbose}{print out prelimianry results such as, which were mapped? Did databases agree?}

\item{genes}{what genes should be in the query?}
}
\description{
Function for transferring a single gene or multiple genes to the consistent gene symbols
}
\note{
ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_info.gz was downloaded on 12DEC2019. This file was converted
into an inverted dict.
}
\examples{
\dontrun{mapp(LFS1)}
}
