% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mkdir}
\alias{mkdir}
\title{make genoppi directories}
\usage{
mkdir(bait, cell, variation = "")
}
\arguments{
\item{bait}{what bait was used?}

\item{cell}{what cell type?}

\item{variation}{add a variation to the outputted name}
}
\value{
will create a directory based on relative paths and return
a list of file names to be used for writing.
}
\description{
Makes a directory according to bait, cell, imputation
}
