% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkhashtable.R
\name{mkhashtable}
\alias{mkhashtable}
\title{Update data.bases}
\usage{
mkhashtable(dat, split = " ", chars.remove = "\\\\*")
}
\description{
A function for generating a hash table that contains the columns 'primary' and 'synonym'
}
