% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderatedTTest.R
\name{moderatedTTest}
\alias{moderatedTTest}
\title{Calculate a moderated t-test}
\usage{
moderatedTTest(df)
}
\arguments{
\item{df}{a data.frame with 'rep' in column names.}
}
\description{
alculate a moderated t-test
}
\author{
April Kim
}
