% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesData.R
\docType{data}
\name{ncbi_aliases}
\alias{ncbi_aliases}
\title{NCBI Hash Table for mapping genes}
\format{An object of class \code{Rcpp_Hashmap} of length 1.}
\usage{
ncbi_aliases
}
\description{
NCBI inverted dictionary hash table that maps an alias to the same gene format.
}
\note{
ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_info.gz was downloaded on 12DEC2019. This file was converted
into an inverted dict. Was downloaded according to this: https://www.biostars.org/p/1378/
}
\author{
Frederik Lassen
}
\keyword{datasets}
