% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandProxies.R
\name{plotProxies}
\alias{plotProxies}
\title{Plot proxies sequentially}
\usage{
plotProxies(data, proxylist, title.vs = "control")
}
\arguments{
\item{data}{a data.frame containing gene, logFC, pvalue and significant.}

\item{proxylist}{a list of data.frames in which the name of the list item correspond
to the prox bait, and the data.frame is all the interactors.}
}
\description{
takes the data and a list of data.frames (generated with \code{expandProxies}),
and generated an overlap plot.
}
