% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Draw Replicate Scatter Plot}
\usage{
plotScatter(df, bait, title = "")
}
\arguments{
\item{df}{something}

\item{bait}{something}
}
\description{
Draws replicate correlation scatter plot(s)
}
\author{
April Kim / Frederik Heymann
}
