% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plot volcano}
\usage{
plotVolcano(
  df,
  bait,
  title = "",
  sub1 = "proteins detected.",
  sub2 = "significant."
)
}
\arguments{
\item{df}{something here}
}
\description{
Volcano plot
}
\author{
April Kim
}
