% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{pipeline for process proteomics data}
\usage{
prepare(bait, infile, cols = NULL, impute = NULL)
}
\arguments{
\item{bait}{a vector containing that contains the bait and cell type that should be matched in a column.}

\item{infile}{the file that contains the raw data, i.e. accession numbers, intensity values, ratios etc.}

\item{cols}{optional manual entry. A vector of columns that are present in the dataset header. Follows 
the format of columns: acession, bait1, mock1, bait2, mock2, bait3, ..}

\item{infiles}{a vector of paths to the data files (expects .csv)}
}
\value{
a table that can be inputted to genoppi
}
\description{
a function that will process ip data
from different baits, files and cells
}
