% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTransform.R
\name{pTransform}
\alias{pTransform}
\title{Transform data}
\usage{
pTransform(table = NULL, type = "log2")
}
\description{
Normalize a table of label or label free data. 
The function will transform all numeric columns.
intensity value.
}
\author{
flassen
}
