% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp2gene.R
\name{snp2gene}
\alias{snp2gene}
\title{map snps to a specific gene}
\usage{
snp2gene(snp, reference = "data/snp_to_gene.RData", verbose = T)
}
\description{
maps a snp to a gene by using a dict as reference
}
\note{
uses \code{null.omit} from misc.R
}
\examples{
snp2gene('rs142920272')
}
\author{
flassen
}
