% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpToGene.R
\name{snpToGene}
\alias{snpToGene}
\title{map snps to a specific gene}
\usage{
snpToGene(snp)
}
\arguments{
\item{snp}{a vector of snps}
}
\description{
maps a snp to a gene by using a dict as reference
}
\note{
uses \code{null.omit} from misc.R
}
\examples{
\dontrun{snpToGene('rs142920272')}
}
\author{
flassen
}
