% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesData.R
\docType{data}
\name{uniprot_aliases}
\alias{uniprot_aliases}
\title{Uniprot for mapping protein names to primary name}
\format{An object of class \code{Rcpp_Hashmap} of length 1.}
\usage{
uniprot_aliases
}
\description{
Uniprot inverted dictionary hash table that maps an alias to the same protein format.
}
\note{
Downloaded 12DEC2019 by https://www.uniprot.org/uniprot/?query=reviewed%3Ayes+AND+proteome%3Aup000005640,
and selected for 'gene names (synonyms) and 'gene names (primary)'. The synonyms provides the keys and the primary
provide the corresponding value.
}
\author{
Frederik Lassen
}
\keyword{datasets}
