% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn}
\alias{warn}
\title{warnings to stderr}
\usage{
warn(msg, file = NULL)
}
\arguments{
\item{msg}{the message.}

\item{file}{write logs to file as well as stderr.}
}
\description{
sends a message to stderr
}
