% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe data}
\usage{
describe(data, control = "mock", verbose = F)
}
\arguments{
\item{data}{a data.frame with proteomics data}

\item{verbose}{print to stdout}
}
\description{
Return boolean vector with length equal to number of columns, 
indicating which index contains intensity, ratios, baits column.
}
\author{
flassen
}
