% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoppi.R
\name{genoppi}
\alias{genoppi}
\title{Genoppi analysis pipeline}
\usage{
genoppi(
  input_file,
  bait_name,
  output_stats_file,
  output_plots_file,
  gene_lists_file = NA,
  fc_cutoff = "both",
  fdr_cutoff = 0.1,
  p_cutoff = NA,
  imp_list_file = NA,
  verbose = T,
  aggregate_info = T,
  debug = F
)
}
\arguments{
\item{input_file}{Genoppi input table}

\item{bait_name}{bait name}

\item{output_stats_file}{output file containing moderated t-test stats (must be a .txt file)}

\item{output_plots_file}{output file containing plots (must be a .pdf file)}

\item{gene_lists_file}{file containing gene list file names, or "NA"}

\item{fc_cutoff}{character "positive", "negative", or "both"}

\item{fdr_cutoff}{numeric value or "NA"}

\item{p_cutoff}{numeric value or "NA"}

\item{imp_list_file}{file containing an overview over which proteins were imputed}
}
\description{
the genoppi combined analysis script. This will run the pipeline
generating correlation plots,
}
\note{
functions modified from Genoppi source code
}
\author{
aprilkim/Yuhanshu/flassen
}
