% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlap.R
\name{plotOverlap}
\alias{plotOverlap}
\title{Draw Overlap Plot}
\usage{
plotOverlap(df, bait, reference, title, drawLabel = T)
}
\arguments{
\item{df}{something}
}
\description{
ne list overlap enrichment test + volcano plot
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{enrichment}()},
\code{\link{genoppi.database}()},
\code{\link{genoppi}()},
\code{\link{mttest}()},
\code{\link{plotScatter}()},
\code{\link{plotVolcano}()}
}
\author{
April Kim
}
\concept{genoppi}
