% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Draw Replicate Scatter Plot}
\usage{
plotScatter(df, bait, title = "")
}
\arguments{
\item{df}{something}

\item{bait}{something}
}
\description{
Draws replicate correlation scatter plot(s)
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{enrichment}()},
\code{\link{genoppi.database}()},
\code{\link{genoppi}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotVolcano}()}
}
\author{
April Kim / Frederik Heymann
}
\concept{genoppi}
