% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Draw a scatter plot of replicates}
\usage{
plotScatter(df, bait, title = "")
}
\arguments{
\item{df}{A data.frame containing replicate columns.}

\item{bait}{The bait that was used. A string.}

\item{title}{Add a title to the ggplot.}
}
\description{
Draws replicate correlation scatter plot(s)
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{enrichment}()},
\code{\link{interactors}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotVolcano}()}
}
\author{
April Kim / Frederik Heymann
}
\concept{genoppi}
