% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plot volcano}
\usage{
plotVolcano(
  df,
  bait,
  title = "",
  sub1 = "proteins detected.",
  sub2 = "significant."
)
}
\arguments{
\item{df}{a data.frame containing columns logFC, significant and p-value.}

\item{bait}{what bait was used?  a string.}

\item{title}{should the ggplot ahve a title?}

\item{sub1}{The first part of a subtitle that is to be generated.}

\item{sub2}{The second part of the subtitle.}
}
\description{
Volcano plot
}
\seealso{
Other genoppi: 
\code{\link{designate}()},
\code{\link{enrichment}()},
\code{\link{genoppi.database}()},
\code{\link{genoppi}()},
\code{\link{mttest}()},
\code{\link{plotOverlap}()},
\code{\link{plotScatter}()}
}
\author{
April Kim
}
\concept{genoppi}
