% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strSplitGene.R
\name{strSplitGene}
\alias{strSplitGene}
\title{splits an accession id}
\usage{
strSplitGene(vec)
}
\description{
splits the gene in a column/vector of accession id
so that the gene is extract.
}
\author{
flassen
}
